+++ resources/js/utils/watermarkHelpers.ts
@@ -1,0 +1,11 @@
+export function needSizeVariantsWatermark(sizeVariants: App.Http.Resources.Models.SizeVariantsResouce): boolean {
+	return (
+		(sizeVariants.thumb && !sizeVariants.thumb.is_watermarked) ||
+		(sizeVariants.thumb2x && !sizeVariants.thumb2x.is_watermarked) ||
+		(sizeVariants.small && !sizeVariants.small.is_watermarked) ||
+		(sizeVariants.small2x && !sizeVariants.small2x.is_watermarked) ||
+		(sizeVariants.medium && !sizeVariants.medium.is_watermarked) ||
+		(sizeVariants.medium2x && !sizeVariants.medium2x.is_watermarked) ||
+		false
+	);
+}
