+++ resources/js/v8/components/gallery/photoModule/Dock.vue
@@ -63,6 +63,7 @@ import { useRoute } from "vue-router";
 import { usePhotoStore } from "@/stores/PhotoState";
 import { useAlbumStore } from "@/stores/AlbumState";
 import { FILL_OVERRIDE_CLASS } from "@/v8/icons";
+import { needSizeVariantsWatermark } from "@/utils/watermarkHelpers";
 
 const toast = useAppToast();
 const lycheeStore = useLycheeStateStore();
@@ -120,18 +121,6 @@ function scanFaces() {
 				life: 3000,
 			});
 		});
-}
-
-function needSizeVariantsWatermark(sizeVariants: App.Http.Resources.Models.SizeVariantsResouce): boolean {
-	return (
-		(sizeVariants.thumb && !sizeVariants.thumb.is_watermarked) ||
-		(sizeVariants.thumb2x && !sizeVariants.thumb2x.is_watermarked) ||
-		(sizeVariants.small && !sizeVariants.small.is_watermarked) ||
-		(sizeVariants.small2x && !sizeVariants.small2x.is_watermarked) ||
-		(sizeVariants.medium && !sizeVariants.medium.is_watermarked) ||
-		(sizeVariants.medium2x && !sizeVariants.medium2x.is_watermarked) ||
-		false
-	);
 }
 
 const emits = defineEmits<{
